+++ src/Plate/index.tsx
@@ -152,13 +152,6 @@ export function convertPositionFromRowToColumnFlow(position: number): number {
   return positionForCoordinates(coordinates, 'column');
 }
 
-const LINE_STYLE = {
-  padding: 4,
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-};
-
 function assertUniquePositions(data: Array<PlateWell>): void {
   const positions = data.map(
     (well) => `${well.coordinates.row}${well.coordinates.column}`,
@@ -198,10 +191,17 @@ export function Plate(props: PlateProps) {
           gridGap: '3px',
         }}
       >
-        <span style={LINE_STYLE} />
+        <span />
 
         {COORDINATES_COLUMNS.map((column) => (
-          <span style={LINE_STYLE} key={column}>
+          <span
+            style={{
+              display: 'flex',
+              justifyContent: 'center',
+              padding: 4,
+            }}
+            key={column}
+          >
             <strong>{column}</strong>
           </span>
         ))}
@@ -228,29 +228,50 @@ export function Plate(props: PlateProps) {
 }
 
 function Well(props: { position: number; well?: PlateWell }) {
+  const generalWellStyle = {
+    backgroundColor: props.well?.color ?? PALETTE.gray3,
+    border: `1px solid ${PALETTE.gray4}`,
+    borderRadius: 2,
+    boxShadow: `0 0.5px 1.5px ${PALETTE.gray4}`,
+    padding: 4,
+  };
+
+  if (props.well?.content) {
+    return (
+      <span
+        style={{
+          ...generalWellStyle,
+        }}
+      >
+        {props.well?.content}
+      </span>
+    );
+  }
   return (
-    <span
+    <small
       style={{
-        backgroundColor: props.well?.color ?? PALETTE.gray3,
-        border: `1px solid ${PALETTE.gray4}`,
-        borderRadius: 2,
-        boxShadow: `0 0.5px 1.5px ${PALETTE.gray4}`,
-        ...LINE_STYLE,
+        color: PALETTE.gray1,
+        display: 'flex',
+        justifyContent: 'center',
+        alignItems: 'center',
+        ...generalWellStyle,
       }}
     >
-      {props.well?.content ?? (
-        <small style={{ color: PALETTE.gray1 }}>
-          {rowForPosition(props.position, PLATE_FLOW) +
-            columnForPosition(props.position, PLATE_FLOW)}
-        </small>
-      )}
-    </span>
+      {rowForPosition(props.position, PLATE_FLOW) +
+        columnForPosition(props.position, PLATE_FLOW)}
+    </small>
   );
 }
 
 function RowLabel(props: { position: number }) {
   return (
-    <span style={LINE_STYLE}>
+    <span
+      style={{
+        display: 'flex',
+        justifyContent: 'center',
+        alignItems: 'center',
+      }}
+    >
       <strong>{rowForPosition(props.position, PLATE_FLOW)}</strong>
     </span>
   );